Extract the finger-tracing engine out of Trace
Behaviour unchanged. This is the refactor CLAUDE.md's modularity rule asks for
before the change that prompted it: a cursive tracing game wants the same
engine, and the alternative was copying four hundred lines into a second file.

Nothing in the waypoint resampling, the hit testing, the pulsing next-dot, the
side columns of controls, the progress bar or the "Great job" badge is specific
to printed letters. It is specific to TRACING. So it moved to LetterTracer,
and TraceGame is now three facts: which glyph table, which alphabets, and its
own name.

Two things had to generalise rather than move verbatim. The alphabet tabs were
three hard-coded rects and a three-valued enum; they are now a table of
{label, first, count}, so a game with two alphabets leaves the third slot empty
instead of relocating Prev -- a control that moves depending on which game you
opened is a control you have to look for. And counts replaced end indices,
because an off-by-one in a boundary silently offers the wrong letter while a
count that is wrong is obviously wrong.

The tracer does not touch the Game's dirty flags. It raises its own and the
shell drains them, because reaching into a protected base from a helper would
make the helper part of the class it exists to be independent of. Both flags
are drained every frame before either is acted on: markFullDirty() raises both,
so a short-circuit would leave the plain one set and spend the next frame on a
repaint nothing asked for.

TraceGame keeps Stroke and Glyph as aliases so TraceGlyphData.cpp -- 260 lines
of hand-authored letterforms -- was not touched by a refactor that changed no
letter in it.
