+++ src/games/TraceGame.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "LetterTracer.h"
 #include "engine/Game.h"
 #include "ui/Ui.h"
 
@@ -7,6 +8,13 @@ struct AppMetadata;
 
 const AppMetadata& traceAppMetadata();
 
+/* Trace printed letters and digits: ABC, abc, 123.
+ *
+ * A shell over LetterTracer, which is where the tracing actually lives. This
+ * class is now only three facts: which glyph table, which alphabets, and its
+ * own name. Everything that used to be here moved when Cursive needed the
+ * same engine -- see LetterTracer.h.
+ */
 class TraceGame : public AppGame {
 public:
     const char* title() const override;
@@ -14,57 +22,12 @@ public:
     void update(AppContext& host, const TouchPoint& touch) override;
     void render(AppContext& host) override;
 
-    struct Stroke {
-        const int16_t* pts;
-        uint8_t count;
-    };
-
-    struct Glyph {
-        char label;
-        const Stroke* strokes;
-        uint8_t strokeCount;
-    };
+    /* The glyph data model, kept here as aliases so TraceGlyphData.cpp -- 260
+     * lines of hand-authored letterforms -- did not have to be touched by a
+     * refactor that changed no letter in it. */
+    using Stroke = LetterTracer::Stroke;
+    using Glyph = LetterTracer::Glyph;
 
 private:
-    static constexpr uint8_t MAX_POINTS  = 96;
-    static constexpr uint8_t MAX_STROKES = 4;
-    static constexpr uint8_t UPPER_FIRST = 0;
-    static constexpr uint8_t LOWER_FIRST = 26;
-    static constexpr uint8_t DIGIT_FIRST = 52;
-    static constexpr uint8_t GLYPH_COUNT_TOTAL = 62;
-
-    enum class GlyphSet { Upper, Lower, Digit };
-
-    struct Pt {
-        int16_t x, y;
-    };
-
-    void loadGlyph();
-    void resampleWaypoints();
-    int16_t scaleX(int16_t nx) const;
-    int16_t scaleY(int16_t ny) const;
-    void drawGuide(Ui::Renderer& tft);
-    void drawProgress(Ui::Renderer& tft);
-    void drawCompleteStatus(Ui::Renderer& tft);
-    void drawModeTabs(Ui::Renderer& tft);
-    void updatePulsePhase();
-    void previousGlyph();
-    void nextGlyph();
-    uint8_t getSetFirstIndex() const;
-    uint8_t getSetLastIndex() const;
-
-    Pt pts_[MAX_POINTS] = {};
-    uint8_t strokeStart_[MAX_STROKES] = {};
-    uint8_t strokeLen_[MAX_STROKES] = {};
-    uint8_t strokeCount_ = 0;
-    uint8_t activeStroke_ = 0;
-    uint8_t nextPoint_ = 0;
-
-    uint8_t glyphIndex_ = 0;
-    GlyphSet glyphSet_ = GlyphSet::Upper;
-    bool complete_ = false;
-    uint32_t completeAt_ = 0;
-
-    uint32_t lastPulseChange_ = 0;
-    bool pulseState_ = false;
+    LetterTracer tracer_;
 };
